+++ src/main/java/com/eventux/backend/repository/AddressRepository.java
@@ -4,7 +4,8 @@ import com.eventux.backend.model.Address;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
+
 @Repository
-public interface AddressRepository extends JpaRepository<Address, Long> {
+public interface AddressRepository extends JpaRepository<Address, Integer> {
 
 }
